+++ vllm/model_executor/models/gpt2.py
@@ -40,7 +40,8 @@ from vllm.model_executor.layers.vocab_parallel_embedding import (
 from vllm.model_executor.model_loader.weight_utils import default_weight_loader
 from vllm.model_executor.sampling_metadata import SamplingMetadata
 from vllm.sequence import IntermediateTensors, SamplerOutput
-from vllm.utils import make_layers
+
+from .utils import make_layers
 
 
 class GPT2Attention(nn.Module):
